added 'populate' controller for ease, testing making a match history entry in new match db table
