+++ prisma/schema.prisma
@@ -73,5 +73,5 @@ model Match {
 	lost			User	@relation("lost", fields: [user2Id], references: [id])			
 }
 // optional if doing more than 2 player matches: participated
-// date			DateTime
+//	date			DateTime
